refactor(kb-connectors): extract canonical-field logic into useConnectorConfigFields hook
